- Fixed bug in SNES checksum calculation code if the cartridge size byte of the   internal header has an invalid value. - If the cartridge size byte is 0, the "internal size" is reported as 0 Mbit   instead of some large value. Additonally, the maximum value is clamped to 128   Mbit. - Added support for a few more cartridge configurations to --ufosd based on   output of --dbuh for dumps made by Mystic_Merlin with his UFO SD. Still more   information is needed. - Made DJGPP port compile again.

+++ src/misc/misc.c
@@ -253,7 +253,7 @@ change_mem (char *buf, unsigned int bufsize, char *searchstr,
   if ((sets = (st_cm_set_t *) malloc (n_esc * sizeof (st_cm_set_t))) == NULL)
     {
       fprintf (stderr, "ERROR: Not enough memory for buffer (%u bytes)\n",
-               n_esc * sizeof (st_cm_set_t));
+               (unsigned int) (n_esc * sizeof (st_cm_set_t)));
       return -1;
     }
   va_start (argptr, offset);
